refactor(parsing): разделены правила нормализации feed
FeedNormalizationService оставлен facade для полного normalize flow
metadata, entry, text/html, URL и icon URL policy вынесены в focused normalizers
